Inventory lazy-loaded from serialized version
